add css

+++ frontend/components/task_modal/index.jsx
@@ -6,6 +6,7 @@ import TaskEditForm from './task_edit_form';
 import style from './index.scss';
 
 import { editTask, deleteTask } from 'actions/task_actions';
+import { disableTaskModal } from 'actions/modal_actions';
 import { taskSelector } from 'reducers/selectors';
 
 const TaskModal = (props) => {
@@ -18,23 +19,39 @@ const TaskModal = (props) => {
     props.deleteTask(props.task.id);
   }
 
-  return (
-    <div className='task-modal-background'>
-      <div className='task-modal-container'>
-        <TaskEditForm
-          onSubmit={handleEditTask}
-          task={props.task}
-        />
+  const handleDisableTaskModal = () => {
+    props.disableTaskModal();
+  }
 
-        <button
-          className="delete-task-button"
-          type="delete"
-          onClick={handleDeleteTask}>
-          delete task
-        </button>
+  if (props.active) {
+    return (
+      <div className='task-modal-background'>
+        <div className='task-modal-container'>
+          <button
+            className="disable-task-modal"
+            type="disable"
+            onClick={handleDisableTaskModal}>
+            x
+          </button>
+          
+          <TaskEditForm
+            onSubmit={handleEditTask}
+            task={props.task}
+            />
+
+          <button
+            className="delete-task-button"
+            type="delete"
+            onClick={handleDeleteTask}>
+            delete task
+          </button>
+
+        </div>
       </div>
-    </div>
-  )
+      )
+  } else {
+    return null;
+  }
 }
 
 const mapStateToProps = state => ({
@@ -45,6 +62,7 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch => ({
   editTask: (task) => dispatch(editTask(task)),
   deleteTask: (id) => dispatch(deleteTask(id)),
+  disableTaskModal: () => dispatch(disableTaskModal()),
 })
 
 export default withRouter(connect(
